Econometric Modeling and Time Series Forecasting
As part of my senior year computer science independent study with Dr.William Verts, I solely developed a business intelligence Python application to input a company’s data and run many regression analyses to find the best linear combination of independent variables (inverse, log, cubic, etc.) with the goal of ultimately modeling supply and demand equations. This would highlight where to produce to maximize revenue, where to price, what the ideal combination of inputs was, how much to spend on things like advertising and much more. 
It accounts for time by adding a trend variable and seasonality dummy variables. It likewise coverts categorical variables to dummy variables and generates interaction terms between the independent variables. In ModelFitter(), the independent variables are regressed against the dependent in various shapes: log, linear, quadratic, cubic, inverse. The shape yielding the highest adjusted R^2 is the optimal relationship between the independent and the dependent. The user would be in control of how confident they want to be in these results, specifying either 50, 80, 90, 92, 95, 98, or 99% confidence in statically significant variables. Having greater confidence, say 95%, would allow you to say “I have 95% confidence that increasing advertising by a dollar increases output by 4.28 units”. However, setting such a high threshold means that other inputs, say the price of a competitor’s product, might be omitted from the regression equation all together because it doesn’t quite meet the correct critical value. Having less confidence, say 80%, would allow more basic information to be revealed about the relationships between the inputs and the output. “With 80% confidence, a $1 increase in a competitor’s item’s price increases output by 5 units”. 5 units may not be the exact number, but it gives you a better understanding of the dynamics between competitor’s price and output, i.e that if the price of my competitor’s product goes up, I can be 80% confident this will increase my own output.

It performs checks for violations of regression assumptions, autocorrelation and heteroskedasticity, and uses a modified version of OLS in these cases. To test for heteroskedasticity it performs White’s Test, which is an auxiliary regression of the model’s variables against the model’s errors square and based on the Chi-squared test determines if the data violate this assumption. If they do, it would ideally perform Feasible Generalized Least Squares (FGLS), but uses GLSAR with one lag to get to a similar result because FGLS is not available in statsmodels package. To test for autocorrelation, a Breusch Godfrey test is performed and based on critical values it may reject the Null hypothesis of no autocorrelation. Again, FGLS would be used to accommodate this or in this case GLSAR.

Once the model was fitted, the program would output easy to understand metrics, such that a nontechnical user could comprehend and make informed business decisions. It would tell the user information like “a one dollar increase in advertising results in a 4.28 unit increase in output.” The user could also play around with varying the amount of an input, because they might incorrectly surmise a constant rate of growth for an input having say, a cubic relationship, rather than a linear one. So the user could ask a question like “if I increase advertising by 50 units, what increase in output would that be?”. 

In addition, to predict future sales it tries some time series forecasting techniques, the Holt-Winters method and Vector Autoregression (VAR) in addition to the previously estimated model and selects the model with the lowest root mean squared error. It tries Holt-Winter’s with an additive or multiplicative trend for annual data and additive or multiplicative seasonal component for higher frequency data. It also performs a Johansen test to check for stationarity.

There is another file in the same folder called portfolioPersonalFinance which includes a section for personal finance questions as well as investment decisions.